feat: uses create(de)Cipheriv instead of deprecated methods
BREAKING CHANGE: min node is 12.14.x, upgrades internal dependencies and hardens minimum
shared secret requirements. Introduces new token format of <version><iv><payload>, so that
in the future breaking changes are easier to mitigate. For back-compatibility introduces
ability to specify legacy secret with a length of 24, which is supported by createCipher/createDecipher.